+++ src/any_singleton/__init__.py
@@ -1,9 +1,10 @@
-from singletons import (
+from .singletons import (
     singleton, singleton_value, singleton_instance,
     once, run_once, SecondCallingBehaviour,
+    SealedObject,
     CannotCallingMoreThanOnceError,
     is_singleton_exists,
-    get_cached_return,
+    get_cached_return
 )
 
 singleton = singleton
@@ -14,6 +15,8 @@ once = once
 run_once = run_once
 
 SecondCallingBehaviour = SecondCallingBehaviour
+
+SealedObject = SealedObject
 
 CannotCallingMoreThanOnceError = CannotCallingMoreThanOnceError