+++ src/any_singleton/singletons.py
@@ -1,10 +1,11 @@
 from enum import Enum
 from functools import wraps
+from types import NoneType
 from typing import Any, Callable, overload
 
 
-class Package:
-    def __init__(self, value: Any) -> None:
+class SealedObject:
+    def __init__(self, value: Any = None) -> None:
         self.__value = value
 
     @property
@@ -14,6 +15,12 @@ class Package:
     @value.setter
     def value(self, value: Any) -> None:
         self.__value = value
+
+
+#class Singleton: pass
+
+
+#class SingletonObject: pass
 
 
 class SingletonInfo:
@@ -49,28 +56,20 @@ def is_singleton_exists(dn: str) -> bool:
     return dn in _g
 
 
-def _is_nonreferenced(obj: Any) -> bool:
-    known_nonreferenced_types = (
-        bool, int, float, str
-    )
-    return isinstance(obj, known_nonreferenced_types)
-
-
 def _add_singleton(dn: str, value: Any) -> None:
-    if _is_nonreferenced(value):
-        obj = Package(value)
-    else:
-        obj = value
-
-    _g[dn] = SingletonInfo(dn, obj)
+    _g[dn] = SingletonInfo(dn, value)
 
 
 def _get_singleton(dn: str) -> Any:
     return _g[dn].instance
 
 
+def _get_return_dn(func_dn: str) -> str:
+    return func_dn + RETURN_POSTFIX
+
+
 def _remove_singleton(dn: str) -> None:
-    return_dn = dn + RETURN_POSTFIX
+    return_dn = _get_return_dn(dn)
     if return_dn in _g:
         del _g[return_dn]
     del _g[dn]
@@ -78,15 +77,15 @@ def _remove_singleton(dn: str) -> None:
 
 def singleton_value(dn: str, value: Any) -> Any:
     """
-    Used for creating a singleton.
+    Creating a singleton with given value.
 
     ATTENTION
-    If you give a non-referenced type value, this function will not seal the value.
-    It means that the value can be changed outside.
+    If you directly give a non-referenced type value, this function will not seal the value.
+    It means that the singleton object will not be changed outside.
 
     :param dn: The unique domain name for the instance.
                It could be like "project.module.variable".
-    :param value: The value in a Non-referenced type to be stored in the singleton.
+    :param value: The default value of the singleton.
     :return: The singleton instance.
     """
 
@@ -100,35 +99,68 @@ def singleton_value(dn: str, value: Any) -> Any:
 
 def singleton_instance(dn: str, t: type, *args, **kwargs) -> Any:
     """
-    Used for creating a singleton of referenced types.
+    Creating an instance as singleton.
+
+    ATTENTION
+    If you directly create an instance of a non-referenced type, this function will not seal the instance.
+    It means that the singleton object will not be changed outside.
 
     :param dn: The unique domain name for the instance.
-               It could be like "project.module.function".
-    :param t: The type of referenced type.
+               It could be like "project.module.variable".
+    :param t: The type of instance to be created.
     :param args: The arguments to be passed to initialize an instance of given type.
     :param kwargs: The keyword arguments to initialize an instance of given type.
+    :return: The singleton instance.
     """
 
     if dn not in _g:
         instance = t(*args, **kwargs)
 
-        _add_singleton(dn, isinstance)
+        _add_singleton(dn, instance)
 
         return instance
     else:
         return _get_singleton(dn)
 
 
-@overload
-def singleton(dn: str, value: Any) -> Any:
-    """
-    Used for creating a singleton of non-referenced types (like int, float, str, etc.).
+_nonreferenced_types = singleton_value(
+    'any_singleton.nonreferenced_types',
+    (bool, int, float, str, NoneType)
+)
 
-    Although it available for giving a value of referenced type, it is not recommended.
+
+def _is_nonreferenced(obj: Any) -> bool:
+    return isinstance(obj, _nonreferenced_types)  # or (obj is type and issubclass(obj, _nonreferenced_types))
+
+
+@overload
+def singleton(dn: str, value: bool | int | float | str | None) -> SealedObject:
+    """
+    Creating a singleton with given value.
+    The function will automatically seal the value into a SealedObject.
 
     :param dn: The unique domain name for the instance.
                It could be like "project.module.variable".
-    :param value: The value in a Non-referenced type to be stored in the singleton.
+    :param value: The default value of the singleton.
+    :return: The singleton instance.
+    """
+    pass
+
+
+@overload
+def singleton(dn: str, value: Any) -> Any:
+    """
+    Creating a singleton with given value.
+
+    ATTENTION
+    If you directly give a non-referenced type value,
+    and it isn't a bool, int, float or str,
+    this function will not seal the value.
+    It means that the singleton object will not be changed outside.
+
+    :param dn: The unique domain name for the instance.
+               It could be like "project.module.variable".
+    :param value: The default value of the singleton.
     :return: The singleton instance.
     """
     pass
@@ -137,13 +169,14 @@ def singleton(dn: str, value: Any) -> Any:
 @overload
 def singleton(dn: str, t: type, *args, **kwargs) -> Any:
     """
-    Used for creating a singleton of referenced types.
+    Creating an instance as singleton.
 
     :param dn: The unique domain name for the instance.
-               It could be like "project.module.function".
-    :param t: The type of referenced type.
+               It could be like "project.module.variable".
+    :param t: The type of instance to be created.
     :param args: The arguments to be passed to initialize an instance of given type.
     :param kwargs: The keyword arguments to initialize an instance of given type.
+    :return: The singleton instance.
     """
     pass
 
@@ -154,25 +187,23 @@ def singleton(dn: str, x: Any, *args, **kwargs):
     if not is_type and (len(args) > 0 or len(kwargs) > 0):
         raise ValueError('Too many arguments.')
 
-    #if dn not in _g:
-    #    if is_type:
-    #        t = x
-    #        v = t(*args, **kwargs)
-    #    else:
-    #        v = x
-    #
-    #    _add_singleton(dn, v)
-    #
-    #    ret = v
-    #else:
-    #    ret = _get_singleton(dn)
-    #
-    #return ret
+    if dn not in _g:
+        if is_type:
+            t = x
+            v = t(*args, **kwargs)
+        else:
+            v = x
 
-    if is_type:
-        return singleton_instance(dn, x, *args, **kwargs)
+        if _is_nonreferenced(v):
+            v = SealedObject(v)
+
+        _add_singleton(dn, v)
+
+        ret = v
     else:
-        return singleton_value(dn, x)
+        ret = _get_singleton(dn)
+
+    return ret
 
 
 class CannotCallingMoreThanOnceError(Exception):
@@ -180,6 +211,10 @@ class CannotCallingMoreThanOnceError(Exception):
 
 
 class SecondCallingBehaviour(Enum):
+    """
+    The behavior when the function is called more than once.
+    """
+
     """
     Returns the result of the first calling.
     """
@@ -196,27 +231,8 @@ class SecondCallingBehaviour(Enum):
     NoneToReturn = None
 
 
-def _add_cached_return(dn: str, value: Any) -> None:
-    """
-    Set the cached return value of a function decorated with `@once`.
-
-    :param dn: The unique domain name of the function.
-    :param value: The value to be cached.
-    """
-
-    singleton_value(dn + RETURN_POSTFIX, value)
-
-
-#def _is_return_exists(dn: str) -> bool:
-#    """
-#    Check if the cached return value of a function decorated with `@once` exists.
-#
-#    :param dn: The unique domain name of the function.
-#    :return: - `True`  if the cached return value exists.
-#             - `False` otherwise.
-#    """
-#
-#    return is_exists(dn + RETURN_POSTFIX)
+def _add_cached_return(dn: str, value: Any) -> Any:
+    return singleton_value(_get_return_dn(dn), value)
 
 
 def get_cached_return(dn: str) -> Any:
@@ -227,10 +243,11 @@ def get_cached_return(dn: str) -> Any:
     you can call the function once again to get the cached return value.
 
     :param dn: The unique domain name of the function.
+               It will not be sealed as SealedObject automatically.
     :return: The cached return value.
     """
 
-    return _get_singleton(dn + RETURN_POSTFIX)
+    return _get_singleton(_get_return_dn(dn))
 
 
 def once(
@@ -246,6 +263,8 @@ def once(
     :return: The decorator.
     """
 
+    return_dn = _get_return_dn(dn)
+
     def decorator(f: Callable) -> Callable:
         """
         A decorator to ensure that the function is only called once.
@@ -253,29 +272,47 @@ def once(
         :return: Decorated function.
         """
 
-        @wraps(f)
-        def wrapper(*args, **kwargs) -> Any:
-            if dn not in _g:
-                _add_singleton(dn, f)
+        if dn in _g:
+            wrapper = _get_singleton(dn)
 
-                ret = f(*args, **kwargs)
+            # if not isinstance(wrapper, Callable):
+            #    raise TypeError(f'"{dn}" has been registered with a non-callable object.')
 
-                if second_calling == SecondCallingBehaviour.CacheReturn:
-                    _add_cached_return(dn, ret)
+            return wrapper
+        else:
+            @wraps(f)
+            def wrapper(*args, **kwargs) -> Any:
+                """
+                :return: The return value or cached value of the function.
+                         It will not be sealed as SealedObject automatically.
+                :except CannotCallingMoreThanOnceError: If the second_calling set with SecondCallingBehaviour.CacheReturn,
+                                                        when the function is called more than once,
+                                                        it will raise CannotCallingMoreThanOnceError.
+                """
 
-                return ret
+                if return_dn not in _g:
+                    ret = f(*args, **kwargs)
 
-            match second_calling:
-                case SecondCallingBehaviour.CacheReturn:
-                    return get_cached_return(dn)
-                case SecondCallingBehaviour.RaiseError:
-                    raise RuntimeError(f'"{dn}" has already been called.')
-                case SecondCallingBehaviour.NoneToReturn:
-                    return None
-                case _:
-                    raise ValueError(f'Unknown "{SecondCallingBehaviour.__name__}".')
+                    if second_calling == SecondCallingBehaviour.CacheReturn:
+                        _add_cached_return(dn, ret)
+                    else:
+                        _add_cached_return(dn, None)
 
-        return wrapper
+                    return ret
+
+                match second_calling:
+                    case SecondCallingBehaviour.CacheReturn:
+                        return get_cached_return(dn)
+                    case SecondCallingBehaviour.RaiseError:
+                        raise CannotCallingMoreThanOnceError(f'"{dn}" has already been called.')
+                    case SecondCallingBehaviour.NoneToReturn:
+                        return None
+                    case _:
+                        raise ValueError(f'Unknown "{SecondCallingBehaviour.__name__}".')
+
+            _add_singleton(dn, wrapper)
+
+            return wrapper
 
     return decorator