+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import pytest
+
+import tests.fixtures.clean_up